requirements.txt
uvicorn
python-dotenv
requests
langchain
langchain-openai
langchain-community
langchain-huggingface
huggingface_hub
chromadb
sentence-transformers
PyMuPDF
tqdm
pydantic
httpx
openai
typing-extensions
colorama
torch